Basic block fading API. Has issues. Details:
1) With Transvoxel, it produces cracks.
This is because when a block fades into a different LOD, neighbors
immediately switch their transition mesh, causing a crack that
remains visible until the fading finishes. To fix this, transition meshes
also need to fade.
2) With shadow mapping, meshes fading out project shadow on their
fading-in counterpart. We need a way to either exclude them from the
shadow map and make the fading-in mesh directly appear in the shadow map,
or doing the effect in post-process somehow.
3) Discarding pixels causes noise depending on where it's done in the shader.
See https://github.com/godotengine/godot/issues/34966