Mesh deallocation seems to not be an issue anymore.
If it still is, it can be moved to a thread since storage is thread-safe.
Profiling has shown however mesh instances can be slow to deallocate now
(details in a comment added in recent commits), but it's not as bad as how
slow meshes used to be. Need to keep an eye on that.