Clamp distant normalmaps to fix artifacts.
On a very distant surface, it is possible that the mesh is so biased that
SDF sampled on its triangles reaches internal features that are normally
not visible (like caves, or the other side of a thin wall). This causes
normals to be flipped. So instead we prevent normals from deviating
further than some maximum angle towards the triangle's normal.