Implemented buffer data reduction in VoxelGeneratorGraph.
Before, a unique buffer was allocated for every node output.
Now buffers are re-used for multiple nodes. Which buffers are re-used is
determined ahead of time (as part of compile()). Some buffers can be
"pinned" to remain unique. Others have fake users so they are re-used
until they "reach" an output.
Dynamic optimization (execution map) should still work, but local constants
required to introduce "constant fills" at runtime, when a skippable node
output is still being used.
Constants and buffers that are re-used for multiple XZ slices are
typically the ones that are pinned. Unconnected node inputs are considered
constants, so we are not getting as much re-uses as we could atm.