Wrap thread-locals in functions to initialize on-demand.
Initialize on demand instead of when the application starts for every
thread. This works around a bug in the static debug MSVC runtime (/MTd).
At time of writing, Godot is using /MT even in debug builds, which
prevents from getting safety checks from the standard library. But when
/MTd is used, global scope static thread_local variables holding standard
library objects make the runtime crash.
See:
https://github.com/baldurk/renderdoc/issues/1743
https://developercommunity.visualstudio.com/t/race-condition-on-g-tss-mutex-with-static-crt/672664