Attempt to workaround issue with new EditorPlugin behavior...
Unfortunately the graph editor is still janky due to the behavior changes
in Godot, and it's unclear what can be done to fix it. The core of the
issue is the fact this editor uses the inspector to edit node properties,
and switches back to the graph resource itself when the background is
clicked. But that causes all sorts of issues... the VisualShaderEditor
somehow gets away with it, but uses stuff that is not exposed to scripts
and extensions.