Fix crash when opening graph resource with a missing function reference.
Also try to not interpret the graph as "new" in the editor if it actually
failed, so don't load default preset.

Godot reports this to the user with a dialog when opening scenes, or when
opening resources for the first time, but not always: if you open a scene
with the missing dependency, it still loads the scene, and if you open
the graph after that, it opens without showing you the dialog again....
so we have to handle the mess ourselves in such cases.