Indicate that child classes of VoxelTool only contain specific methods.
Some users didn't understand why some child classes had little to no API.
It's because only functions specific to a voxel container will be listed,
while all functions inherited from VoxelTool are also available.
And it can be empty because while the API is the same, the implementation
in the engine can be different.