Changes to enforce the intented locking patterns in VoxelData.
This is to prepare safe access to it from multiple threads. Very few
features do that yet, but more will in the future.
Initially done because I was going to use it directly to store partially
generated chunks as they get processed, but later I decided to investigate
using a separate data structure, so I'm not actually using this.
Still, it may be useful for other cases.

However, it is untested. All demos need to be profiled again, to check if
this has a particular impact or has introduced bugs.