Introduce explicit task cancellation tokens.
They are more reliable than a loosely updated viewers position list. The
scheduler can decide to cancel tasks explicitely, and the task
doesn't have to know about the logic of such cancellation.
Only downside at the moment is that it relies on a shared_ptr, which means
memory allocation and an extra fat pointer to store in tasks.