Added a check to remove degenerate/microscopic triangles from Transvoxel.
This is documented in the Transvoxel paper. We ignored it so far because
it is quite rare (less than 1% of meshes in typical terrain) and did not
cause problems in practice... so far.

This is to workaround errors thrown by Jolt Physics, which doesn't ignore
them instead. Microscopic triangles can also cause meshes to become empty
and cause further errors, due to Jolt having to re-index meshes passed
from the Godot Physics3DServer API.

Unfortunately, that change makes meshing 15% slower (texturing off).