Fix incorrect task prioritization when spawning first viewer and terrain.
If both were added at the same frame in a specific order, VoxelTerrain
could tick before VoxelEngine due to how it relies on the
VoxelEngineUpdater node. So all tasks spawned by the viewer in the
terrain's first update would get a shared priority dependency that didnt
include the viewer, causing tasks to be prioritized around world origin.

The new code workarounds this corner case by preallocating entries into
the shared dependency and using an atomic count, so viewers can be added
without invalidating the dependency. So a stockpile of tasks will get
sorted correctly eventually. There is a limit to this fix though, if there
are a lot of viewers. We may find a better strategy if a game needs to use
many.