Cache neighbor pointers in data blocks to speedup box queries.
This is not suitable when data blocks are sparse. It assumes
a block is present everywhere that is loaded, so probably works with
data streaming to check if all blocks are present in an area. But it
doesn't work if the goal is to actually retrieve these blocks in sparse
storage scenarios (like storing only edited ones, or not streaming).