WIP enforce I/O task execution order per chunk.
This is because the threaded task runner doesn't guarantee execution order
of tasks it is given. Things usually works because tasks are spaced
enough in time and to not intertwine... unless the runner becomes
overloaded (too many expensive tasks at once). In that case, it is possible
for saves and loads to run in different order than their scheduling order.
Task priority is only indicative and cannot enforce this.

So we have to queue subsequent tasks somehow, while keeping the benefits
of priority sorting (run earlier if closer to players).
Issues occur within the scope of chunks. Different chunks
can be considered independent in this context. So we could associate a
queue to every chunk with pending tasks.

I'm not sure yet if we'll have to come to this. Also, it might still not
cover all corner cases. Perhaps this has to be extended to more tasks, not
just I/O. More solutions might need to be designed, for example chunk
versionning to handle responses in the right order too.
Instancer is annoyingly difficult to apply this to, because it can use
different chunk sizes.