Introduced alias of std::vector using our own default allocator.
Therefore it will use `memalloc` and `memfree` when compiling as a Godot
module or GDExtension, instead of directly using the C++ runtime's
default. It should be equivalent to the original default, except now
allocations will be tracked in Godot's counters.

TODO A bunch of areas weren't changed because of our usage of static
thread_local. I don't know yet if it is safe to let these objects be
destroyed automatically while using Godot, considering GDExtensions are
dynamic libraries, and Godot's allocating functions are passed when the
library initializes, rather than being known to be from the C++ runtime.

In parallel to this, it would be interesting to find alternatives to the
`static thread_local` pattern, because it's being used a lot now and it's
probably not a good idea to keep doing it this way.