Changed remaining std::vectors to StdVector.
As explained in the previous commit, they were left like this because I
wasn't sure of destruction safety in GDExtension.

According to the standard, `thread_local` objects should be destroyed when
the thread completes. We complete our threads upon unregistration of the
library, which does destroy the objects.

There is still some degree of doubt. For example, depending on how you use
the standard library, threads might not actually end:
https://developercommunity.visualstudio.com/t/thread-local/798234
However that seems to only apply to the parallel algorithms library, which
we don't use. We use Godot threads, which are based on basic threads, so
they should complete when we expect them to.

In the event this still somehow doesn't hold true, some ideas:
- Revert this commit
- Encapsulate all thread_local objects into a "TLSWrapper" object that
  would track initialized ones in a global list, and explicitely
  destroy them by iterating that list when the library is unregistered
- Get rid of thread_locals somehow, for example using thread index from our
  own pool (this approach still needs experimentation, users could call
  our APIs from threads our library hasn't created)