Add debug option to workaround a way to see the navmesh in the editor.
NavigationRegions don't show up in the editor if they are not owned by
the edited scene. That means we can't have any preview if our plugin
automatically generates regions. So one workaround is to have the option
to temporarily set the owner of regions. Unfortunately that's not enough,
because Godot doesn't update the scene tree when owners change. It requires
switching to another scene tab and back.