Implement deferred merges instead of merging immediately.
This is meant to fix flickering that occurs when outdated meshes from the
parent LOD are updated lazily, but slightly too late since they are
activated immediately instead of waiting for them to be ready.

However that seems to bring new issues as LOD fading flickers badly again
when moving backwards.