Initial implementation for partial side culling. Only works with quads.
The goal was to make it work using some kind of 2D mesh boolean
subtraction, however I could not find an easy way to do it so far. Godot
exposes some methods in Geometry2D based on Clipper, but they don't seem
to be enough when the input is an arbitrary mesh.
So for now I implemented the rest, and constrained the culling part to
only work on quads. If a side isn't a quad, it will fallback to render
the full side like before.
This is also a relatively expensive feature for niche use cases, compared
to other potential alternatives, so it may remain experimental for now.