Fix multi-materials on VoxelLodTerrain (not full support).
This allows using VoxelMesherBlocky materials.
If a non-ShaderMaterial is set on the terrain, it acts as override.
If a ShaderMaterial is assigned on the terrain, things might not work.
Multi-material requires that we also detect whether each of them is
a ShaderMaterial in order to support LOD features such as:
- Per block info
- Transition masks
- Detail textures
- Fading
When using the blocky mesher, not all of this is useful, so it won't be
trivial to implement this in an optimal way. Per-instance uniforms
are still an option eventually, but not every renderer supports them and
they seem to require adjustment (buffer size) otherwise they throw errors