WIP refactor GPU code to defer all RenderingDevice usage to its own thread.
Doesn't compile. Had to jump temporarily on something else...

This is because of a change introduced in Godot 4.4 (dev), not yet
announced or documented, where some methods of RD can only be used in the
thread it was created in (they have guard macros).
This made our current code spam errors because we created our RD on
startup before we start our custom GPU thread, and we have some code
creating resources outside of that thread. Most of functions to create
resources can actually be used outside the thread, but a few things can't,
including freeing them.
So the current attempt is to post everything as a task and let it
execute on the thread in the submission order.