Workaround vox scene importer lockup.
Apparently, saving meshes deadlocks the editor (4.3 and 4.4 dev).
Importing runs in a thread by default, so I suspect it is a
synchronization issue. Forcing to run on the main thread fixes it,
but I don't know if this is a bug or a known thing to do when writing
scene importers.