Use new `cache_limit` SCons parameter.
This is not only to use an up-to-date API: if that limit isn't set,
Godot's build scripts will print something with the character '∞'
(infinite) in it. This unfortunately leads Python to throw an exception
on the Windows CI machine due to incorrect encoding, which in turn
makes the job fail.
I tried to workaround that issue with PYTHONIOENCODING earlier, but for
some reason it didn't have any effect.
See https://github.com/godotengine/godot/issues/103520