Fix cases where Godot did auto-free.
In this task there are a mix of uniform sets that Godot frees and doesn't
free automatically (had to test it to confirm which ones...).
Initially I thought of adding another vector in debug builds to verify
that the uniform sets we expect to be auto-freed are actually freed,
but a simpler approach seems to be to free them all before everything else.