WIP voxel_size property. Untested.
Acts like another scale factor, but doesn't affect child nodes. It may
however affect chunk size, which in turn affects mesh density, which in
turn may affect VoxelInstancer.
Mesh data itself is not scaled, instead voxel_size is included in the
transform. So we have to account for that in VoxelInstanceGenerator.