Added side panel to the graph editor to workaround inspector issues.
When we inspect sub-resources of nodes (like Noise, Curve etc), they
open in the inspector. But if the graph is used in another resource,
that means we loose track of it, and we no longer see its preview.
Workaround: graphs have a list of "users" that they can register to.
When the graph editor edits a new graph and finds a user, it may add
a side panel to display that user, with the option to (re)open it in the
inspector.